Admin All Products &Create Product Page Task Done

+++ backend/routes/ItemRoute.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const { getAllItems, createItem, updateItem, deleteItem, getItemDetail, itemReview, getItemsReviews, deleteItemReview } = require("../controllers/ItemController");
+const { getAllItems, createItem, updateItem, deleteItem, getItemDetail, itemReview, getItemsReviews, deleteItemReview, getAdminAllItems } = require("../controllers/ItemController");
 const {isAuthenticatedUser,authorizeRoles} = require("../middleware/auth");
 const router = express.Router();
 
@@ -9,6 +9,7 @@ router.route("/admin/item/create")
 router.route("/admin/item/:id")
 .put(isAuthenticatedUser,authorizeRoles("admin"),updateItem)
 .delete(isAuthenticatedUser,authorizeRoles("admin"),deleteItem)
+router.route("/admin/allItems").get(isAuthenticatedUser,authorizeRoles("admin"),getAdminAllItems);
 
 router.route("/item/:id").get(getItemDetail);
 router.route("/review").put(isAuthenticatedUser,itemReview);
